Create Notes from Tony
Overall, solid job :) A couple of minor comments... Challenge 1 - Step 2 - regarding the question, "What level is the reference level?" you responded, 0.75971, which is the beta coefficient for Migration level 2, but the reference level is "Migration level 1". I think you are a bit confused about what I was asking for. - Step 4 - you calculate both the original F stat p value (calculated by the aov() function, evaluating the test statistic is relative to an F distribution with 3 and 11000 degrees of freedom, which is what you print out) AND by permutation, which is your variable p.value, which you don't print out (which is "zero"). These are (rightly) very close to one another and essentially zero; what I was asking you for was the value calculated by permutation Challenge 2 - Step 3 - you don't plot the Range Size distribution, either raw or log transformed (and the latter should be used for the subsequent steps). Here, too, you seem a bit confused with what I am asking for with respect to the reference levels/the outcome of the Tukey test. Regarding the question "which Migration categor(ies) are different than the reference level?" you say "3-2", and regarding the question "What level is the reference level?" you say 780559.2... those don't make sense. Migration style levels "2" *and* "3" are different from the reference level ("1"), and "2" and "3" are also different from each other. - Step 4 - you make boxplots and run linear models for the full dataset, not for passerines only, despite filtering to get a new passerines data frame; as a result, while the approach is correct, you don't generate the correct answers for this step - Step 6 - regarding the question, "Based on the model output, what would you now conclude about how relative beak length is related to these two variables?" you write: "They can be predicted by both as well" This is cursory and not complete... based on this model, rel beak length is significantly predicted by primary lifestyle, tropic level, and the interaction between these two predictors.